fix: component name & mobile calendar css
