+++ src/styles/globals.scss
@@ -9,3 +9,12 @@ body {
   line-height: normal;
   font-family: hiragino-kaku-gothic-pron, sans-serif;
 }
+
+:root {
+  --fc-button-bg-color: transparent;
+  --fc-button-border-color: transparent;
+  --fc-button-hover-bg-color: transparent;
+  --fc-button-hover-border-color: transparent;
+  --fc-button-active-bg-color: transparent;
+  --fc-button-active-border-color: transparent;
+}
