+++ src/ui/MobileCalendar/index.module.scss
@@ -2,4 +2,4 @@
 
 .container {
     border: 12px double #FFC785;
-  }
+}
